refactor: 날짜 계산 관련 코드 모듈화
